+++ locales/ruRU.lua
@@ -52,4 +52,5 @@ L["DEATHKNIGHT_USEPATHOFFROST"] = "Использовать %s при вызов
 L["DEATHKNIGHT_USEONLYINWATERWALKLOCATION"] = "Использовать только в зонах с хождением по воде"
 L["SHAMAN_USEWATERWALKING"] = "Использовать %s при вызове наземного средства передвижения"
 L["SHAMAN_USEONLYINWATERWALKLOCATION"] = "Использовать только в зонах с хождением по воде"
+L["DRUID_USELASTDRUIDFORM"] = "Вернуть последнюю форму при спешивании"
 L["DRUID_USEMACROALWAYS"] = "Использовать этот макрос вместо средств передвижения"